refactor: move download logic from tools/ to assets/download.py
Per @awsarron review (PR #84 threads on download.py and download_assets.py):
tools should be lightweight wrappers that expose strands-robots library
functions to agents, not contain the actual implementation.

Before:
  tools/download_assets.py (493 lines) — all download logic + @tool
  assets/download.py (14 lines) — just re-exports from tools

After:
  assets/download.py (421 lines) — all download logic (library code)
  tools/download_assets.py (78 lines) — thin @tool wrapper

Changes:
- assets/download.py: moved download_robots(), all backends
  (_download_via_robot_descriptions, _download_via_git,
  _download_from_github), and helpers from tools/download_assets.py
- tools/download_assets.py: slim @tool wrapper that parses action/args
  and delegates to assets.download.download_robots()
- assets/manager.py: fix _auto_download_robot() lazy import from
  tools.download_assets → .download (same package, relative import)
- tools/__init__.py: add download_assets to lazy imports

No circular imports: assets/download.py imports from assets/manager.py
at module level; manager.py lazy-imports from assets/download.py inside
_auto_download_robot() only.

Tests: 238 passed, 6 skipped, 0 failures.
